STYLE: Añadido componente SectionListTaskView con DisclosureGroup, TaskViewComponents
